requirements.txt
amqp==5.3.1
asgiref==3.9.1
async-timeout==5.0.1
attrs==25.3.0
billiard==4.2.1
boto3==1.39.14
botocore==1.39.14
celery==5.5.3
certifi==2025.7.14
channels==4.2.2
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
contourpy==1.3.2
cycler==0.12.1
Django==5.2.4
djangorestframework==3.16.0
django-filter==25.1
fonttools==4.59.0
idna==3.10
jmespath==1.0.1
kafka-python==2.2.15
kiwisolver==1.4.8
kombu==5.5.4
matplotlib==3.10.3
narwhals==1.48.1
numpy==2.2.6
opencv-python==4.10.0.84
packaging==25.0
pandas==2.3.1
pillow==11.3.0
plotly==6.2.0
projectaria-tools==1.6.0
prompt_toolkit==3.0.51
psycopg2-binary==2.9.10
pyarrow==21.0.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
redis==6.2.0
channels-redis==4.3.0
daphne==4.1.2
django-redis==5.4.0
requests==2.32.4
rerun-sdk==0.24.0
s3transfer==0.13.1
six==1.17.0
sqlparse==0.5.3
tqdm==4.67.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.13
